add helpers.hpp, fix leaks
